pyproject.toml
[project]
name = "struct-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jaro-winkler>=2.0.3",
]
